Move out of usecases package
